db정보추가/integer변경

+++ src/main/java/com/example/gym/repository/GymRepository.java
@@ -3,5 +3,5 @@ package com.example.gym.repository;
 import com.example.gym.domain.Gym;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GymRepository extends JpaRepository<Gym, Long> {
+public interface GymRepository extends JpaRepository<Gym, Integer> {
 }
