+++ src/main/java/com/example/gym/repository/MemberRepository.java
@@ -3,5 +3,5 @@ package com.example.gym.repository;
 import com.example.gym.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberRepository extends JpaRepository<Member, Long> {
+public interface MemberRepository extends JpaRepository<Member, Integer> {
 }
